requirements.txt
fastapi==0.116.1
python-multipart==0.0.21
uvicorn[standard]==0.35.0
supabase==2.10.0
pypdf2==3.0.1
python-docx==1.1.2
pandas==2.2.2
python-dotenv==1.2.1
pytest==8.3.3
google-generativeai==0.8.6
fpdf2==2.8.2
